Move the treeDataGridAdapter to the ViewLoadoutGroupFilesViewModel for consistency with other view models
